+++ .golangci.yml
@@ -1,0 +1,26 @@
+linters:
+  enable-all: true
+
+  disable:
+    - gomnd
+    - nosnakecase
+    - varnamelen
+    - ifshort
+    - exhaustivestruct
+    - golint
+    - deadcode
+    - structcheck
+    - varcheck
+    - interfacer
+    - maligned
+    - scopelint
+    - exhaustruct
+    - rowserrcheck
+    - wastedassign
+    - exhaustive
+    - ireturn
+
+
+run:
+  timeout: 10m
+  tests: false
