+++ go.mod
@@ -1,0 +1,5 @@
+module github.com/AlexZav1327/guess-game
+
+go 1.20
+
+require github.com/go-telegram-bot-api/telegram-bot-api/v5 v5.5.1
